requirements.txt
beautifulsoup4
faiss-cpu
fastapi
googlemaps
jinja2
langchain-community
langchain-openai
numpy
openai
passlib
pyinstaller
pydantic
python-dotenv
python-jose
requests
tensorboard
torch
transformers
ttkthemes
typing_extensions
uvicorn
